% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AAspectra.R
\docType{class}
\name{AAspectra-class}
\alias{AAspectra-class}
\title{AAspectra S4 object}
\value{

}
\description{
AAspectra S4 object
}
\section{Slots}{

\describe{
\item{\code{spectra}}{data.frame.}

\item{\code{charge}}{numeric.}

\item{\code{AAstr}}{character.}
}}

\examples{
AAsOBJ <- new("AAspectra",
spectra = spectra,
charge = charge,
AAstr = AAstr)
}
