% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AAspectra.R
\name{getAAspectra}
\alias{getAAspectra}
\title{Get AAspectra object from AA sequence with natural SIP abundance}
\usage{
getAAspectra(AAstr, charge = 1)
}
\arguments{
\item{AAstr}{amino acide string}

\item{charge}{charge of the ion}
}
\value{
AAspectra object
}
\description{
Get AAspectra object from AA sequence with natural SIP abundance
}
\examples{
getAAspectra("KHRIP",2)
}
