% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AAspectra.R
\name{getSipPrecursorSpectra}
\alias{getSipPrecursorSpectra}
\title{Get AAspectra object  of precursor from AA sequence
with labeled SIP abundance}
\usage{
getSipPrecursorSpectra(AAstr, Atom = "C13", Prob = 0.01, charge = 2)
}
\arguments{
\item{AAstr}{amino acide string}

\item{Atom}{a CharacterVector C13 or N15}

\item{Prob}{C13 or N15's abundance}
}
\value{
AAspectra object
}
\description{
Get AAspectra object  of precursor from AA sequence
with labeled SIP abundance
}
\examples{
getSipPrecursorSpectra("KHRIPCDRK", "C13", 0.05, 2)
}
