% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{peak_calculator}
\alias{peak_calculator}
\title{Simple peak calculator of natural isotopic distribution}
\usage{
peak_calculator(AAstr)
}
\arguments{
\item{AAstr}{a CharacterVector}
}
\description{
Simple peak calculator of natural isotopic distribution
}
