% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readScansMS1}
\alias{readScansMS1}
\title{readScansMS1}
\usage{
readScansMS1(ftFile, scanCount)
}
\arguments{
\item{ftFile}{a ft1 file's full path}

\item{scanNumber}{the scanNumber th scan}
}
\description{
readScansMS1
}
